update `AppRoutes` type
